+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * `attr::style`
 * `attr::minlength_u16`
 * `attr::maxlength_u16`
+* `attr::role`
 
 
 ### Documentation